add ref local reassignment
And build project around the sample.
