+++ snippets/csharp/programming-guide/ref-returns/Program.cs
@@ -1,0 +1,18 @@
+// #error version
+
+namespace UseRefReturns
+{
+    using static System.Console;
+    using RefReturns;
+    using RefReturnsVersionTwo;
+
+    public class Example
+    {   
+       static void Main(string[] args)
+       {
+            FirstExample.Tests();
+            WriteLine();
+            SecondExample.Tests();
+       }
+    }
+}
