+++ snippets/csharp/programming-guide/ref-returns/ref-returns.csproj
@@ -1,0 +1,10 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp2.0</TargetFramework>
+    <LangVersion>7.3</LangVersion>
+    <RootNamespace>RefReturns</RootNamespace>
+  </PropertyGroup>
+
+</Project>
